Refactor ExecutorServiceBuilder to separate package & tweak properties
