+++ core/server/common/src/test/java/alluxio/master/ExecutorServiceBuilderTest.java
@@ -11,7 +11,7 @@
 
 package alluxio.master;
 
-import alluxio.ExecutorServiceBuilder;
+import alluxio.executor.ExecutorServiceBuilder;
 import alluxio.conf.PropertyKey;
 import alluxio.conf.ServerConfiguration;
