+++ core/server/master/src/main/java/alluxio/master/AlluxioMasterProcess.java
@@ -14,7 +14,7 @@ package alluxio.master;
 import static alluxio.util.network.NetworkAddressUtils.ServiceType;
 
 import alluxio.AlluxioURI;
-import alluxio.ExecutorServiceBuilder;
+import alluxio.executor.ExecutorServiceBuilder;
 import alluxio.RuntimeConstants;
 import alluxio.conf.PropertyKey;
 import alluxio.conf.ServerConfiguration;
